refactor: work 도메인 엔티티들 패키지 변경
